add site title
Text::Template does not seem to like blessed objects, another reason why
it is not the right template system for this thing...

We also need to make sure to test that the correct content is present,
and not just compare pages against each other. If the template has an
error, testing page renders against each other will not reveal the
problem, since they'll still be equal.

Fixes #19