ignore blog directories that aren't YYYY/MM/DD
This allows me to put a /drafts folder at least and not have it picked
up by the blog app. In the future, we'll make a way to manage the
drafts, but for now let's allow any directory that doesn't look like
what we want it to be.

We may in the future support just YYYY/MM, but we will probably never
support just YYYY.