add coercions for Theme and Store
Since these objects are mostly just paths, why not coerce the real
object from the simple path string? It makes the initial setup a lot
easier.

The only tradeoff is that it's not as obvious what is happening in the
site config file, so it may discourage others from creating their own
Theme or Store.

Fixes #69