fix --help not working
Mojolicious::Commands has to set some things up at startup, which may
come from arguments in @ARGV, so it runs GetOptions on @ARGV. Since one
of the tested options is "help|h", it removes that option from @ARGV
before we get to it.

For now, we save @ARGV before we load anything else, to make sure that
nobody can take something out of it. In the future, we might change
Mojolicious to not do this (if we can, it does it for a good reason) or
we can also fix this by moving our mojo app to another module which gets
loaded at runtime after we've done all our @ARGV stuff.

Fixes: #106

Thanks to @tadegenban for all the help tracking this down!