make sitemap.xml a template
Individual documents now have a changefreq of "weekly". Every page has a
changefreq and lastmod set (previously, list pages did not set lastmod).

The sitemap.xml template is currently only in the default theme, with no
way to set what theme gets used. We may want to set a site-wide theme
that gets used by default when creating apps.

Feed pages continue to not show up in the sitemap, because they are not
to be indexed.

Fixes #115