paginated list pages should share last_modified
The last modified of all members of a paginated list should be the last
modified of the most recent document in that list, since every page has
to be modified in order to push everything down in the list.

last_modified needs to be accurate because we give it to search engines
for them to reindex our site.

Fixes #125