add test_constructor helper function
The test_constructor helper allows me to quickly test that required
attributes are required and default attributes are set to something
resembling the right value.

In the future I want a better declarative syntax for expressing the
coercions that should be available, but for right now this saves a bunch
of typing.

Fixes #132