reset STDIN to the tty when piping in content
If we do not do this, Vim warns "Vim: Warning: Input is not from a
terminal", and when Vim exits, it leaves our terminal in an unusable
state (how could it not? it doesn't know what terminal we're using!)

This same trick won't work on Windows, so we will need to check to see
if that's actually a problem there too.