require frontmatter to begin with '---'
In order to have frontmatter, the first line of the file must start with
'---'. This will make it easier for us to support frontmatter in other
files in the future, and maybe on a per-section basis...

This is another breaking change, but all the documents produced by
Statocles have always had that first '---'. Only files written from
scratch would not have them.