cache file store's realpath to fix race condition
If our store's directory disappears before we can clean ourselves up (if
it's a temporary directory and the tempdir gets deleted before the
store), trying to find the realpath will throw an exception.

Additionally, when we're in global destruction, we don't care about the
store locations anymore. This prevents spurious warnings about
uninitialized values in hashes if our path object got cleaned up before
we did.