fix RSS pubDate incorrectly using locale setting
The RSS spec requires that pubDate use RFC822 dates, which provide only
for the "C" locale. strftime() is one of the few Perl subs that use the
locale by default (without the `use locale` pragma).

I don't know what this will do on Windows. And I don't know what this
will do if the user doesn't have the locale I switch to (ru_RU), so I
wrapped that in an eval.

Fixes #185