fix tags list not appearing on blog list pages
Apps need to be able to cache pages so that templates can query the apps for
information, like the list of tags. It's more convenient and cleaner to
put these kinds of things on the App and not in the Page (otherwise, the
Page would have to do everything that every app could do).

We may need a way to do arbitrary metadata on both Page and Document so
that we don't have to do this caching.

Fixes #184