move theme to site, removing it from all apps
There is now only a single, site-wide Theme. The ability to have a
single theme per application was not very useful and made some future
development a lot harder.

With a single, site-wide theme, we can add a special /theme URL that
holds theme scripts, stylesheets, and images. Prior to this, we had to
get everything from a CDN, or add a static app to serve the theme's
content (though that solution is fragile at best).

Fixes #126
Refs #175