add new blog posts to store so auto-build works
Previously we used a temporary directory to store the new blog post.
Now, we write the blog post into the same directory where it will live
when the user is done editing. This makes any running `daemon` pick the
blog post up right away so we can edit our post and see the changes
immediately.

Fixes #183

The new file also has a ".markdown" extention, as it should.

Fixes #180