fix unable to set date for blog post from command
Previously, when using "blog post --date YYYY-MM-DD", the warning
"Unknown option: date" was printed, and the date was not used. This was
because Getopt::Long by default cleans out all the things that look like
options that it does not recognize.

Adding "pass_through" to the Getopt::Long configuration fixes this.

Fixes #198