add base_url to Store for per-deploy base URLs
This will allow us to add multiple deploy stores in the future.

We kept the base_url on the Site in the case of simple sites with only
one deploy store. Setting the base_url on a Store requires us to do a
bit more work configuring it, since we can't just do the magic string
coercion.

Fixes #117