add deploy class to deploy a site
The deploy class does the work of the former Statocles::Site::Git
without requiring an entire subclass of Statocles::Site. This opens up
the possibility of other ways of deploying being very simple to set up,
and allowing a single site to have multiple configured deploys (for
development and production, for example).

Refs #63