explicitly require ".ep" when including templates
Previously, `include "something.html"` would check for
"something.html.ep" first, processing that as a template, and only if
that did not exist, try to do "something.html" as a plain file. This
kind of hidden magic is hard to see and fix if something is going wrong.

While writing the Theme help file, I realized this was too much magic.