remove deploy path from the daemon
This fixes the daemon trying to deploy the site.

We used to use the deploy store to get some of the static files from the
deploy directory. We did this because we didn't have the Static app. Now
that we do, users have to use it in order to get the Static capability.