fix static app builds hidden directories like .git
Previously, if you've got a git repository and added a Static app at the
root directory, it would try to open up all of the git objects,
eventually reaching the open file descriptor limit (which is its own
bug).

Now, if any part of the path is hidden, the Static app does not use it.