do not deploy ignored files or submodule files
Before we deploy to Git, we check `git status` to see if git notices the
file, and we pass the `--ignore-submodules` flag to make sure we don't
try to add something from our submodules.

This may have unintended consequences, but this should probably be the
default. It will at least hide some potential issues with copying files
to the build directory that we actually want to completely ignore.

Fixes #229
Refs #181