change page published to last_modified
Now last_modified is required for all pages. For Document and List
pages, it is a proxy to the underlying document/page. For others, it
defaults to "now", but any value may be supplied.

This is a lot cleaner than what we were doing, but it does break any
current theme...

Fixes #130