reorganize guides and add Statocles::Help
In order to make a better crumbtrail, I added Statocles::Help with links
to all the other guides. This means that the guides don't need to be
linked on the main module page anymore, and we might end up changing the
main module page to be something more technical.

Fixes #253