upgrade to Mojolicious 6.0
Mojo::Reactor::is_readable was deprecated 5 days ago and removed.
Luckily, we don't appear to need it. (kraih/mojo@83a05c99)

To fix an issue with using parens in template expressions, a unary + was
added, which means a slight adjustment on our side (kraih/mojo@da152136)

Mojo::DOM->type was changed to match the DOM API (in which "type"
describes the type of element, not the tag name) (kraih/mojo@993416fd).
But, the test using that method was bad anyway, so it doesn't use it
anymore.