add LinkCheck plugin to check links and images
The LinkCheck plugin attaches to the build event and checks to make sure
that all the internal site content is linked correctly.

This has to rebuild all the pages inside of itself though. We need to
give the Pages event the full page content so we don't have to do
that. This also means that the content is not rewritten for the base
URL, so that could hide a problem.

Fixes #78