add -p <port> option to specify port in daemon
If we want to run multiple daemons, or if another mojo app is running,
using the default Mojo port will not work. For now, just allow
specifying a port. In the future, we can allow specifying a full listen
line, if someone somehow wants to use the daemon for external serving.

Fixes #219