use document templates to override page templates
When a document has a template or layout set, the page will use that
instead of the template/layout set by the application.

This has revealed a bunch of technical debt around creating and giving
template _objects_ to the Page. Since the Page has the Site which has
the Theme, we don't need to give it objects. We can just give it
arrays/strings and it can figure it out.

Fixes #40