fix link check showing schema-less urls as broken
A schema-less URL starts with two slashes instead of one, and has the
host and path, but not the schema (http/https). This is useful if a site
is being served over both http and https, or if you just don't want to
care about which it is being served over.

This patch fixes the LinkCheck plugin to not report schema-less URLs as
broken, since they are on another site.

Fixes #310