add warning if app creates duplicate pages
If an app creates two pages with the same path, or if an index.html is
somehow overwritten by the site's "index" attribute, give the user a
warning so they know what's happening.

This also adds a test app that allows testing the site without using one
of the existing apps (to be continued in #319).

Fixes #331