make find_files skip documents by default
There are two different things to do when making pages: Turn all
documents into Statocles::Page::Document, and turn all files into
Statocles::Page::File. By default, neither should interfere with the
other.

The real rationale for this is that we need to be able to distinguish
the two. It's just convienient to be able to say that we don't want
documents while we're looking for files.

Refs #332