fix mojo ioloop tests when testing file events
Firstly it is important that you use Mojo::IOLoop class methods and not
using the singleton directly, because some reactors (EV) have to use a
non-singleton when recursing.

Secondly running tests in a callback is always frought with peril, so in
this case, you don't have to! Just start the loop and let the app notice
the change.

Fixes #353