refer to templates by path for more flexibility
Previously, we only allowed "category" and "template", forcing all
templates in the theme to be "foo/bar". Now, we support any level of
depth, allowing for better organization in themes. This is especially
important because every document can have its own template, potentially
ballooning the number of templates in a site.

Fixes #313