move includes to theme and add caching
Since every single page in the default themes have at least 3 includes,
this should drastically speed up a full site rebuild.

This also forces all includes to be treated as templates, running them
through the template processor.

This is a breaking change that is not going through the deprecation
policy in this pre-v1.0 development.

Fixes #377
Fixes #378