add date option to build blog pages from the future
The app `pages()` method can now take a hash of options. The blog uses
this to have a date option to build the blog as of a date other than
today. This way, we can test the future blog post we are building.

This does not yet have a way of accessing from the command-line yet.
This also doesn't apply to the `recent_posts` method. We may need a
better way to do this...

Refs #246