only set version during test if no version exists
We should be testing our deprecations are being removed when we say they
will. To that end, our tests should know what version they're testing,
so we can make the tests fail if we don't remove our deprecated things.

This means that the only versioned tests will be accurate for
deprecations, but that is what it is.