deprecate write_document method return value
The write_document method of Statocles::Store returns the full path. It
was thought at the time that behavior would provide some encapsulation
to allow for non-file stores. With the removal of
Statocles::Store::File, this is no longer necessary, and removing it
makes things more consistent and more obvious (because a developer must
explicitly access the store's path to get the full path).

Fixes #382