add class frontmatter for custom document classes
The default Statocles document is extremely generic. This makes it good
for a wide variety of uses, but not very good at doing a lot of
customization (like an iTunes podcast), or doing a lot of validation of
things.

By specifying the `class` key in the frontmatter, you can override the
class with a custom one, enabling more frontmatter keys, custom
validations, custom content generation, and more.

This is probably going to be a niche feature, as most of the possible
customizations likely require custom themes as well.

Fixes #48