add disqus blocks to the default themes
This allows users to add a Disqus ID to their site and enable Disqus
comments on their blog.

We could add this more granularly, even to all pages in the site by
manipulating App data or something. We should also have a way to turn
off comments on specific things, but this is fine for now.

We did not write any tests though, which is not fine. We need to have a
way of testing optional additions to the default themes.

Fixes #109