allow writing most page attributes
The entire point of the event hooks, and now applications too, is that
pages can be edited before they are rendered. This does open the user up
to some confusing possibilities, wondering where page changes are coming
from, but the benefit in flexibility is worth it.

You still can't change which app a page came from, or which site.
Everything else is most likely fair game.