refactor Blog app to use Store app role
This ensures that everything using stores is doing it the same way.

We aren't allowing the blog to have pages outside the /YYYY/MM/DD/
directory structure though. We may need to turn that into its own role,
or otherwise make a special case for ignored but visible directories,
like a leading _ or something...

Fixes #343