make Theme into an App
Previously, we added theme files to the site in a custom manner. This
prevents the Theme object from having any control over what files made
it into the site.

Now, the Theme object has control over what goes in, so it can decide
not to add the template files, for example.

Refs #399