do not deploy template files with the site
Previously, the actual templates (`.ep` files) would be deployed with
the rest of the site, which could cause private, or valuable information
to be leaked.

Since templates allow any Perl code, and since the Perl code could
contain sensitive information like passwords or API keys, deploying the
templates will reveal that information.

The templates themselves are also valuable: It would make it easy to
steal the user's theme and use it on other Statocles sites.

Now, the theme explicitly filters the `.ep` files so they are not
deployed with the rest of the site.

Fixes #399