fix inner links do not work in perldoc app
Because of the URL rewriting we were doing, the internal links to
sections of the module documentation were incorrect. We weren't properly
looking for the `#` URL fragment marker to properly append it to the end
of the rewritten URL.

Now, section links are created correctly, with the fragment marker on
the end.

Fixes #366