switch to warn that a store path doesn't exist
Previously, we would die if a store path doesn't exist. This creates
a chicken-egg problem: If we want to run `statocles bundle theme` to
bundle a default theme, but the theme's store directory doesn't exist,
we can't create the site object in order to bundle the theme to create
the theme directory...

    # This will die with an error
    $ rm -rf theme
    $ statocles bundle theme default

In general, it isn't that big of a deal if a store directory doesn't
exist. It just means there are no documents/files for that store. So,
now we simply warn the user that it doesn't exist, and then we move on
our merry way.

Fixes #414
Fixes #421