escape document title to fix special characters
We should be able to have `<` and `>` in the page title, so make sure to
escape it.

Interestingly, Mojo::DOM automatically unescapes things, so we aren't
testing to see if things got escaped, we're testing to see that things
weren't interpreted as HTML. If these tests fail, they will likely fail
with the gotten value being the empty string.

Fixes #422