allow adding links to pages
This makes it easy to add scripts and stylesheets to a page as it is
worked on by plugins.

This is only appending links to the list for now. We may want, in the
future, to remove links, but that's in the future. We can already edit
links by getting the link object and modifying it in-place (if the
Statocles::Link object allows it).