fix highlight breaking with Markdown code blocks
When doing a Markdown code block to add the <pre> and <code> tags
automatically, instead the entire highlighted block would be escaped,
making the highlighting useless.

Also, it's just annoying to have to remember the
`<pre><code>...</code></pre>` wrapper to support a very limited use-case
of allowing inline `<code>...</code>` blocks to be highlighted. If we
want that in the future, we can add an argument to the `highlight`
function, or even make it a new helper.

Fixes #427