fix no order in sitemap causing spurious commits
Since we run our pages through a hash to ensure we don't have
duplicates, they could be listed in any order. If the order changed, the
sitemap would change, causing a deploy commit when no actual content was
being changed.

Now we sort the sitemap by URL, so that only if something actually changes
does the sitemap file get changed.

Refs #430