add child page scripts and styles to the list page
When a child page has a special stylesheet, like a syntax highlighting
color sheet, it's likely required to display that child page correctly.
If a list page fails to include that stylesheet, the list page might not
display that part of the child page correctly, making the list page look
bad.

Fixes #429