show load error if failed to load Pod::Weaver
In cases where the prereqs for Pod::Weaver are messed up, such as
a Clone module that doesn't load, we need to see the actual error
message to get to the source of the problem. Just saying that
"Pod::Weaver is missing" is not helpful, especially when we literally
just installed it.

Fixes #440