fix travis always having certain prereqs installed
When you specify a two-part Perl version, travis picks one of its
distributions of Perl. These distributions helpfully install a bunch of
prereqs for you. This can hide bugs in your own prereqs though, if one
of the installed prereqs matches one that you need but have not
declared.

By specifying a 3-part version, we get a clean copy, and can be assured
that our prereqs are being checked.

Also, by getting a pristing 5.10, we can test for ancient versions of
core modules, and ensure that we're specifying the right prereq
versions.