allow templates to be overridden for each app
Now, each app has a `templates` attribute which is a map of name -> path
pairs. Users can use this attribute to override the default template
with one of their own.

All apps now also have a `template` method which will get the template
with the given name. The default for a name is generated from the app's
class. Now, applications can have predictable template names for similar
features. For example, if multiple applications need RSS feeds, they can
share the "index.rss" name, so any plugin that works on RSS feeds knows
the RSS feed template for that application is called "index.rss".

Refs #312