rename moniker to template_dir and make explicit
"template_dir" is a better description of what this attribute does at
the moment.

Making it explicit and not based on the object's package will help when
roles are composed into applications, which can change the package name.
This also makes it more clear where templates are coming from, and makes
small inheritance tasks a bit easier.

Refs #312