switch to DateTime::Moonpig for date/time functions
The Time::Piece module has a problem with parsing dates before
1900-01-01. Since it'd be nice for document metadata to accurately
reflect the content inside, we need a module that will support a wide
range of dates.

DateTime::Moonpig is Perl's venerable DateTime module with a lot of the
footguns (such as mutability) removed.

This is a breaking change for templates: The RSS template used the
`tzoffset` method, which does not exist in DateTime::Moonpig.

Fixes #461