always push the git deploy
This fixes a problem where a deploy gets interrupted after the commit
but before the push completes successfully. In this case, just running
"statocles deploy" isn't enough to update the site anymore. You have to
check out the deploy branch, push the changes to the right remote and
branch, and then check out the main branch again.

By pushing every time, we can make this case easier to handle.

Fixes #470