add better error message when weaver.ini is missing
When using the Perldoc app together with Pod::Weaver, there must be
a weaver.ini file. Without it, Pod::Weaver will die with an error "no
viable configuration could be found" and a stack trace as long as my
arm. Out of context, this error message is incomprehensible.

Now we check for the file expected to be there and throw a friendlier
error if it does not exist.

Fixes #446