move content sections to page object
The template state hash was problematic. The state had to be cleared for
each page, because templates are cached and used with multiple pages.

But more to the point, the state did not belong there. The page is
generating this content, so the page should keep hold of it. Then it
doesn't need to be cleared out at all.