allow bundling of specific files from a theme
With the bundle theme command, we can now grab specific files from the
default theme. This makes it easier to build a theme from scratch,
especially for those files that don't really change much.

This is preferable to simply defaulting to the default theme if
a template is missing, so as to avoid spooky behavior.