pass-through the document data attr to the page
This makes it more uniform during theme creation: We always access the
`$page` object when we want content from the document.

This could create a conflict if an application wants to add new data but
also allow the document's data. When that time comes, we could create an
application-only hash of extra template variables.