warn about links with empty href destination
I often make blank links as a note to myself to remember to fill in
a URL. If I forget, though, there will be a link that just brings the
user to the top of the page, which isn't what I want. So now, we warn on
these empty links. If you want to bring the user to the top of the page,
use an explicit `#` character.

Fixes #492