fix section 2 link in feed with only one section
If there's only one section, the section 2 link in the feed will go to
the top of the page. But the link text is "Continue reading...", which
is kind of confusing.

Now we only show this link if there is another section. And finally
we've got some tests for the feed content.

Fixes #496