fix newer/older button being links with no href
When we don't have an older or newer page, the button to go to the
older/newer page must not be clickable. Previously, we could click the
button and go to a place, whether it be the top of the current page
(bad) or a "Not Found" page (worse).

This gets a bit into the differences between the paginators in the two
frameworks we use for our default themes. Let's hope this doesn't open
up a can of worms here...