use blog post date field to sort posts in a day
If a blog has more than one post in a day, the order they appear in is
alphabetical by their path. This can make posts that are written later
appear earlier, which means readers might not see a post.

For now, we're fixing this by using the `date` field in the post to work
out ties. Later, we'll allow configuring how the path/slug is generated
so that authors can adjust things themselves (including adding an simple
counter for posts written in the same time period).

Fixes #512