fix html lint messages to contain filename
The HTML::Lint module will report the filename it knows as long as you
give it a file name. Since this also ensures that the stream is
fully-parsed and processed, it is going to be required to do this in the
future. So, we're getting the jump on this to fix two problems.

Thanks @petdance for explaining the proper use of the HTML::Lint API