add "-raw" flag to prevent render of includes
When we're including a file that contains content that _looks_ like
a Mojolicious template directive, we get strange errors that are
difficult to track down. We should fix this by requiring people to use
certain extensions to enable rendering (like `.ep` and the document
extensions as known by the store (`.markdown` and `.md` by default)),
but we should also allow people to override these heuristics.

So, for now, we'll add the flag that allows overriding the behavior, and
then later we can fix our heuristics for determining if we should run
the content through the renderer.

Refs #529