add non-filesystem based storage backend; update tests and Store to work with it
A new storage backend, Store::Archive::Tar is included as an example
and to allow testing.

In order to accomodate testing of multiple backends the t/store
tests were converted into Moo::Role's.  These are applied to a
My::Test::Store class which is instantiated for each storage backend.
The roles are discovered via Mojo::Loader, so it's simple to add new
ones.  The disadvantage to this approach is that the tests are no
longer easily parallalizeable.

The t/store test suite assumed filesystem based storage, and often
went directly to disk rather than use the Store API.  The suite now
uses the Store API exclusively. There was one addition to the API to
support this, namely Store::read_file_raw(), which unlink read_file()
peforms no decoding of the data from the file.