add disable_content_template attribute
This attribute can be set on documents, applications, and sites to
enable/disable content template processing.

This isn't an ideal solution, but it is an expedient one. In the future,
we might make the content templates more configurable (allowing users to
set which characters introduce template directives, or which template
processor to use, for example).

Refs #541