move build/deploy code into commands
The site object is now responsible for generating / rewriting pages, and
the commands are responsible for writing that content somewhere. We're
trying to reduce the amount of filesystem writing we need to do,
especially in tests.

This means changes to the deploy API to pass in pages to the deploy()
method, not a store object.

The build command also now takes a directory to build to.