allow configuring template directive characters
When we're writing blog posts about Perl, we run afoul of the template
parser, a lot. The `%` that indicates a line of template is the same as
the hash sigil, which often appears on the start of a line.

Now, users can disable line-based processing entirely if they want, and
change the template tags to look however they want (maybe to match
Template Toolkit, so `[% ... %]`).

Refs #541