Avoid can't call method "log" on undefined object
When creating a new Statocles project, if you use the default theme, the
code throws an exception, trying to call site->log.

site is undefined at this point. A better fix might be to initialize it
first, but I wasn't sure if site can exist before the Statocles project
has been created.