move default build directory into the store
Previously, the build directory was based on the current directory. This
would effectively be the same directory, except during the tests. So,
the mojo app test was failing unless there was a `.statocles/build`
directory in the current directory that it could use.

Using the store as a root directory is what I intended and works better.
Having this default copied in multiple places is not good, and should be
addressed.