add build_date to site, set during build
Before this commit any functionality that depended on knowing the build
date (which is possibly overridden by the command line) would need to
have the date passed in to it. With this change the site now knows what
the build date is and is therefore accessible by anything with access to
the site object.