start version of Statocles as a Yancy app
This version of Statocles will read all the static files for the site
into a Yancy database. Then, the export command will export the entire
site, and the deploy command will deploy it.

This drastically simplifies this project, and make it more flexible too,
by relying directly on Mojolicious's and Yancy's features instead of
writing our own.