fix indexes not linking or being linked correctly
When we're serving an index file, the trailing slash needs to be in the
URL for any relative paths to work. So, to make that work right, we need
to redirect from a directory to directory with a slash at the end.

We also need to make sure to remove /index from the end of URLs when
Mojolicious's built-in `url_for` would add them. Mojolicious is
technically correct, but URLs look nicer without that rigamarole.