fix database pages not being found
This would, no matter what, use the filesystem to determine if a page
should or should not have a redirect to a URL with a trailing slash.
This would change the ID being searched for and make it harder to find
the database page (therefore defaulting to the filesystem, which has
nothing). The filesystem should be a fallback after the database is
checked.